requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
streamlit>=1.25.0
python-dotenv>=1.0.0
requests>=2.28.0
pydantic>=2.0.0
aiohttp>=3.8.0
markdown>=3.4.0